# openapi.yaml
openapi: 3.0.3
info:
  title: AI数字名片 API
  description: AI数字名片后端服务 - 模块化架构
  version: 2.0.0
  contact:
    name: AI数字名片团队
servers:
- url: http://localhost:8201
  description: 本地开发
- url: https://api.example.com
  description: 生产环境
paths:
...
  /api/trust/network:
    get:
      summary: 获取我的信任网络
      tags:
      - 信任网络
      parameters: []
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
    post:
      summary: 添加信任关系
      tags:
      - 信任网络
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustResponse'
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrustCreate'
  /api/trust/network/{trusted_user_id}:
    delete:
      summary: 移除信任关系
      tags:
      - 信任网络
      parameters:
      - name: trusted_user_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
  /api/trust/network/{user_id}:
    get:
      summary: 获取指定用户信任网络
      tags:
      - 信任网络
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Successful Response
...
  /api/auth/sso/login:
    get:
      summary: 发起SSO登录(302重定向)
      tags:
      - SSO 认证
      parameters:
      - name: provider
        in: query
        required: false
        schema:
          type: string
          default: google
      - name: redirect_after
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Successful Response
  /api/auth/sso/callback:
    get:
      summary: SSO回调
      tags:
      - SSO 认证
      parameters:
      - name: code
        in: query
        required: false
        schema:
          type: string
      - name: state
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Successful Response
...
  /api/auth/sso/link:
    post:
      summary: 绑定SSO账号
      tags:
      - SSO 认证
      parameters: []
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
...
  /api/v1/admin/tenants:
    post:
      summary: 创建租户(admin)
      tags:
      - Tenant 管理
      parameters: []
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
    get:
      summary: 租户列表(admin)
      tags:
      - Tenant 管理
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          default: 20
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
  /api/v1/tenant/profile:
    get:
      summary: 当前租户信息
      tags:
      - Tenant 管理
      parameters: []
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT令牌, 通过 /api/auth/login 获取
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key, 通过 /api/api-keys 创建
  schemas:
...
    TrustCreate:
      type: object
      required:
      - trusted_user_id
      properties:
        trusted_user_id:
          type: integer
    TrustResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        trusted_user_id:
          type: integer
        created_at:
          type: string
          format: date-time
...